feat: Devide Component Of Main Task
