+++ client/index.html
@@ -131,6 +131,7 @@
         <h3 class="task-heading">Tasks</h3>
         <button class="remove-completed-tasks-btn">
           <i class="fas fa-trash-alt"></i>
+          Clear Finished
         </button>
       </div>
       <div class="msg-container">
